requirements.txt
flask==2.0.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==1.0
joblib==1.1.0
werkzeug==2.0.1
python-dotenv==0.19.1
